[tune] Move trainable_kwargs generation to trial.py
