+++ python/ray/tune/tests/test_api.py
@@ -31,7 +31,7 @@ from ray.tune.callback import Callback
 from ray.tune.experiment import Experiment
 from ray.tune.trainable import wrap_function
 from ray.tune.logger import Logger, LegacyLoggerCallback
-from ray.tune.execution.ray_trial_executor import _noop_logger_creator
+from ray.tune.experiment.trial import _noop_logger_creator
 from ray.tune.result import (
     TIMESTEPS_TOTAL,
     DONE,
